fix errors

+++ src/main/java/ru/kata/spring/boot_security/demo/service/UserServiceImpl.java
@@ -1,12 +1,14 @@
 package ru.kata.spring.boot_security.demo.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.kata.spring.boot_security.demo.model.User;
 import ru.kata.spring.boot_security.demo.repository.UserRepository;
 
+import javax.persistence.EntityNotFoundException;
 import java.util.List;
 import java.util.Optional;
 
@@ -41,19 +43,29 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public User showUserById(int id) {
-        Optional<User> userById=userRepository.findById(id);
-        return userById.orElse(null);
+        Optional<User> userById = userRepository.findById(id);
+        return userById.orElseThrow(EntityNotFoundException::new);
     }
+
     @Transactional
     @Override
     public void updateUserById(int id, User updateUser) {
-        updateUser.setId(id);
-        updateUser.setPassword(passwordEncoder.encode(updateUser.getPassword()));
-        userRepository.save(updateUser);
+        if (userRepository.existsById(id)) {
+            updateUser.setId(id);
+            updateUser.setPassword(passwordEncoder.encode(updateUser.getPassword()));
+            userRepository.save(updateUser);
+        } else {
+            throw new UsernameNotFoundException("User not found");
+        }
     }
+
     @Transactional
     @Override
     public void deleteUserById(int id) {
-        userRepository.deleteById(id);
+        if (userRepository.existsById(id)){
+            userRepository.deleteById(id);
+        } else {
+            throw new UsernameNotFoundException("User not found");
+        }
     }
 }
